Añadido el metodo "delete" en el controlador
Se modifico el metodo "post" para q devuelva una respeusta, ademas se añadio la logica necesaria para el metodo "delete" en el Service

+++ src/main/java/com/incade/GestorPP/Repositorio/MovimientoRepositorio.java
@@ -5,6 +5,6 @@ import com.incade.GestorPP.Entidad.Movimiento;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovimientoRepositorio extends JpaRepository<Movimiento, Long> {
+public interface MovimientoRepositorio extends JpaRepository<Movimiento, Integer> {
     List<Movimiento> findByTipo(String tipo);
 }
